security apply

+++ app/api/posters/delete/route.ts
@@ -9,24 +9,42 @@ export async function POST(request: Request) {
     return NextResponse.json({ status: 401, message: "User not logged in" });
   }
 
-  // TODO: Check if user who is deleting the poster is the owner of the poster
-
   try {
     const { posterId } = await request.json();
 
-    await prisma.like.deleteMany({
-      where: {
-        posterId,
-      },
-    });
-
-    const data = await prisma.poster.delete({
+    const poster = await prisma.poster.findUnique({
       where: {
         id: posterId,
       },
     });
 
-    return NextResponse.json({ status: 204 });
+    if (!poster) {
+      return NextResponse.json({
+        status: 404,
+        message: "Poster not found",
+      });
+    }
+
+    if (session.user.role === "ADMIN" || session.user.id === poster.userId) {
+      await prisma.like.deleteMany({
+        where: {
+          posterId,
+        },
+      });
+
+      await prisma.poster.delete({
+        where: {
+          id: posterId,
+        },
+      });
+
+      return NextResponse.json({ status: 204 });
+    } else {
+      return NextResponse.json({
+        status: 401,
+        message: "Not enough permission to do this",
+      });
+    }
   } catch {
     return NextResponse.json({ status: 500, message: "Internal Server Error" });
   }
